requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
slowapi
email-validator
cloudinary
python-multipart
